chore: move extensions to dfx-core
