+++ src/dfx-core/Cargo.toml
@@ -12,6 +12,7 @@ aes-gcm.workspace = true
 argon2.workspace = true
 bip32 = "0.4.0"
 byte-unit = { workspace = true, features = ["serde"] }
+bytes.workspace = true
 candid = { workspace = true, features = ["random"] }
 dialoguer = "0.10.0"
 directories-next.workspace = true
@@ -24,6 +25,7 @@ ic-identity-hsm = { workspace = true }
 k256 = { version = "0.11.4", features = ["pem"] }
 keyring.workspace = true
 lazy_static.workspace = true
+reqwest.workspace = true
 ring.workspace = true
 schemars.workspace = true
 sec1 = { workspace = true, features = ["std"] }
@@ -32,6 +34,7 @@ serde.workspace = true
 serde_json.workspace = true
 slog = { workspace = true, features = ["max_level_trace"] }
 tar.workspace = true
+tempfile.workspace = true
 thiserror.workspace = true
 tiny-bip39 = "1.0.0"
 url.workspace = true
