+++ src/dfx-core/src/extension/manifest/extension.rs
