+++ src/dfx-core/src/extension/manifest/mod.rs
