+++ src/dfx/Cargo.toml
@@ -36,7 +36,7 @@ atty = "0.2.13"
 backoff.workspace = true
 base64.workspace = true
 byte-unit = { workspace = true, features = ["serde"] }
-bytes = "1.2.1"
+bytes.workspace = true
 candid = { workspace = true, features = ["random"] }
 clap = { workspace = true, features = ["derive", "env", "unstable-styles", "wrap_help"] }
 console = "0.15.0"
