+++ src/dfx/src/lib/environment.rs
@@ -1,8 +1,6 @@
 use crate::config::cache::DiskBasedCache;
 use crate::config::dfx_version;
-use crate::lib::error::extension::ExtensionError;
 use crate::lib::error::DfxResult;
-use crate::lib::extension::manager::ExtensionManager;
 use crate::lib::progress_bar::ProgressBar;
 use crate::lib::warning::{is_warning_disabled, DfxWarning::MainnetPlainTextIdentity};
 use dfx_core::config::cache::Cache;
@@ -10,7 +8,9 @@ use dfx_core::config::model::canister_id_store::CanisterIdStore;
 use dfx_core::config::model::dfinity::{Config, NetworksConfig};
 use dfx_core::config::model::network_descriptor::NetworkDescriptor;
 use dfx_core::error::canister_id_store::CanisterIdStoreError;
+use dfx_core::error::extension::ExtensionError;
 use dfx_core::error::identity::IdentityError;
+use dfx_core::extension::manager::ExtensionManager;
 use dfx_core::identity::identity_manager::IdentityManager;
 
 use anyhow::{anyhow, Context};
@@ -255,7 +255,7 @@ impl Environment for EnvironmentImpl {
     }
 
     fn new_extension_manager(&self) -> Result<ExtensionManager, ExtensionError> {
-        ExtensionManager::new(self)
+        ExtensionManager::new(self.get_version())
     }
 }
 
@@ -371,7 +371,7 @@ impl<'a> Environment for AgentEnvironment<'a> {
     }
 
     fn new_extension_manager(&self) -> Result<ExtensionManager, ExtensionError> {
-        ExtensionManager::new(self.backend)
+        ExtensionManager::new(self.backend.get_version())
     }
 }
